Add files via upload
This is the source code that me team 'Terror Bytes' worked on in completing the Entelect University Cup Hackathon. We decided to code the solution in Java and we had a lot of fun figuring out a few things that were new to us - including trying to upload this repository to GitHub!